Add NowFn to placement options
