+++ placement/placement_test.go
@@ -22,7 +22,6 @@ package placement
 
 import (
 	"fmt"
-	"math"
 	"sort"
 	"testing"
 
@@ -438,121 +437,6 @@ func TestVersion(t *testing.T) {
 
 	p1 = p1.SetVersion(100)
 	assert.Equal(t, 100, p1.GetVersion())
-}
-
-func TestMarkShardSuccess(t *testing.T) {
-	i1 := NewEmptyInstance("i1", "", "", "endpoint", 1)
-	i1.Shards().Add(shard.NewShard(1).SetState(shard.Initializing))
-	i1.Shards().Add(shard.NewShard(2).SetState(shard.Initializing).SetSourceID("i2"))
-
-	i2 := NewEmptyInstance("i2", "", "", "endpoint", 1)
-	i2.Shards().Add(shard.NewShard(1).SetState(shard.Initializing))
-	i2.Shards().Add(shard.NewShard(2).SetState(shard.Leaving))
-
-	p := NewPlacement().
-		SetInstances([]Instance{i1, i2}).
-		SetShards([]uint32{1, 2}).
-		SetReplicaFactor(2)
-
-	_, err := MarkShardAvailable(p, "i1", 1)
-	assert.NoError(t, err)
-
-	_, err = MarkShardAvailable(p, "i1", 2)
-	assert.NoError(t, err)
-}
-
-func TestMarkShardFailure(t *testing.T) {
-	i1 := NewEmptyInstance("i1", "", "", "endpoint", 1)
-	i1.Shards().Add(shard.NewShard(1).SetState(shard.Available))
-	i1.Shards().Add(shard.NewShard(2).SetState(shard.Available))
-
-	i2 := NewEmptyInstance("i2", "", "", "endpoint", 1)
-	i2.Shards().Add(shard.NewShard(1).SetState(shard.Initializing).SetSourceID("i3"))
-	i2.Shards().Add(shard.NewShard(2).SetState(shard.Initializing).SetSourceID("i1"))
-	i2.Shards().Add(shard.NewShard(3).SetState(shard.Initializing).SetSourceID("i1"))
-
-	p := NewPlacement().
-		SetInstances([]Instance{i1, i2}).
-		SetShards([]uint32{1, 2}).
-		SetReplicaFactor(2)
-
-	_, err := MarkShardAvailable(p, "i3", 1)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "does not exist in placement")
-
-	_, err = MarkShardAvailable(p, "i1", 3)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "does not exist in instance")
-
-	_, err = MarkShardAvailable(p, "i1", 1)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "not in Initializing state")
-
-	_, err = MarkShardAvailable(p, "i2", 1)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "does not exist in placement")
-
-	_, err = MarkShardAvailable(p, "i2", 3)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "does not exist in source instance")
-
-	_, err = MarkShardAvailable(p, "i2", 2)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "not leaving instance")
-}
-
-func TestMarkShardWithCutoverInFuture(t *testing.T) {
-	timeInFuture := int64(math.MaxInt64)
-	i1 := NewEmptyInstance("i1", "", "", "e1", 1)
-	i1.Shards().Add(shard.NewShard(0).SetState(shard.Leaving).SetCutoffNanos(timeInFuture))
-
-	i2 := NewEmptyInstance("i2", "", "", "e2", 1)
-	i2.Shards().Add(shard.NewShard(0).SetState(shard.Initializing).SetSourceID("i1").SetCutoverNanos(timeInFuture))
-
-	p := NewPlacement().
-		SetInstances([]Instance{i1, i2}).
-		SetShards([]uint32{0}).
-		SetReplicaFactor(1)
-
-	_, err := MarkShardAvailable(p, "i2", 0)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "is scheduled after now")
-}
-
-func TestRemoveInstanceFromArray(t *testing.T) {
-	instances := []Instance{
-		NewEmptyInstance("i1", "", "", "endpoint", 1),
-		NewEmptyInstance("i2", "", "", "endpoint", 1),
-	}
-
-	assert.Equal(t, instances, RemoveInstanceFromList(instances, "not_exist"))
-	assert.Equal(t, []Instance{NewEmptyInstance("i2", "", "", "endpoint", 1)}, RemoveInstanceFromList(instances, "i1"))
-}
-
-func TestMarkAllAsAvailable(t *testing.T) {
-	i1 := NewEmptyInstance("i1", "", "", "endpoint", 1)
-	i1.Shards().Add(shard.NewShard(1).SetState(shard.Initializing))
-	i1.Shards().Add(shard.NewShard(2).SetState(shard.Initializing))
-
-	i2 := NewEmptyInstance("i2", "", "", "endpoint", 1)
-	i2.Shards().Add(shard.NewShard(1).SetState(shard.Initializing))
-	i2.Shards().Add(shard.NewShard(2).SetState(shard.Initializing))
-
-	p := NewPlacement().
-		SetInstances([]Instance{i1, i2}).
-		SetShards([]uint32{1, 2}).
-		SetReplicaFactor(2)
-
-	_, err := MarkAllShardsAsAvailable(p, true)
-	assert.NoError(t, err)
-
-	i2.Shards().Add(shard.NewShard(3).SetState(shard.Initializing).SetSourceID("i3"))
-	p = NewPlacement().
-		SetInstances([]Instance{i1, i2}).
-		SetShards([]uint32{1, 2}).
-		SetReplicaFactor(2)
-	_, err = MarkAllShardsAsAvailable(p, true)
-	assert.Contains(t, err.Error(), "does not exist in placement")
 }
 
 func TestConvertBetweenProtoAndPlacement(t *testing.T) {
